+++ example/execution/abs.c.bak
@@ -1,0 +1,10 @@
+int calc_abs(int num) {
+  int abs;
+  if (num > 0) {
+    abs = num;
+    return abs;
+  } else {
+    abs = num * (-1);
+    return abs;
+  }
+}
