+++ example/execution/greet.c
@@ -1,0 +1,12 @@
+#include <stdio.h>
+
+void greet() {
+/* Start of aspect */
+printf("こんにちは\n");
+/* End of aspect */
+}
+
+int main() {
+  greet();
+  return 0;
+}
