+++ example/execution/greet.c.bak
@@ -1,0 +1,11 @@
+#include <stdio.h>
+
+void greet() {
+  printf("Hello\n");
+  return;
+}
+
+int main() {
+  greet();
+  return 0;
+}
