+++ example/grammar/constructor/foo.js
@@ -1,0 +1,7 @@
+const foo = (func) => {
+  func();
+};
+
+foo(() => {
+  console.log("foo");
+});
