+++ example/grammar/method/abs.acc
@@ -1,0 +1,23 @@
+abstract aspect AbstractFuncAspect {
+  constructor(
+    filename,
+    os_service_id,
+    location,
+    set_flag(),
+    msg,
+    ret,
+    req_sending_service,
+    get_return,
+  )
+  
+  abstract void after_sending_msg()
+
+  after(): set(StatusType task_location) && infile(this.filename) {
+    if (tsk_location == this.location) {
+      this.set_flag()
+      this.req_sending_service(glbtskid_store, this.os_service_id, this.msg);
+      this.ret = (UINT8)(this.get_return value());
+      this.after_sending_msg()
+    }
+  }
+}
