+++ example/grammar/method/conc.acc
@@ -1,0 +1,19 @@
+aspect ConcreteFuncAspect extends AbstractFuncAspect {
+  super(
+    "ActivateTask_common_aspect.c",
+    OSServiceId_ActivateTask,
+    CALL_MULTI,
+    MULTI_SET_FLG(),
+    0xFFFFFFFF,
+    ercd,
+    request_sending_multi_service,
+    get_multi_return,
+  )
+
+  void after_sending_msg() {
+    if (ercd != E_OK) {
+      _errorhook_par1.tskid = TSKID(tskid_store);
+      call_errorhook(ercd, OSServiceId_ActivateTask);
+    }
+  }
+}
